requirements.txt
python-telegram-bot==20.6
langchain==0.0.340
openai==1.3.5
aiohttp==3.9.1
python-dotenv==1.0.0
ipython==8.18.1
mypy==1.7.1
flake8==6.1.0
black==23.11.0